Add conventional test for codehilite markdown extension
